Add files via upload
Done with Pre-Processing for subjects 1-4 from experiment 1 and subjects 6 and 8 from experiment 2(our 6 total subjects). If you run this notebook locally on your own computer, the filtered data won't be correct unless you specify the channels to discard through the interactive plots(before PSD plots). For that reason, reach out to me for the filtered data for our subjects.